requirements.txt
asgiref==3.11.0
crispy-bootstrap5==2025.6
Django==5.2.8
django-ckeditor==6.7.3
django-crispy-forms==2.5
django-js-asset==3.1.2
django-taggit==6.1.0
pillow==12.0.0
python-decouple==3.8
sqlparse==0.5.3
tzdata==2025.2
whitenoise==6.11.0
